pyproject.toml
[tool.poetry]
name = "squigglyinfinitecompilers"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.6.2"
discord = "^1.0.1"
praw = "^7.1.0"
redditcli = "^2.1.0"
youtube-dl = "^2020.7.28"
redset = "^0.5.1"
flask = "^1.1.2"
"disease.py" = "^1.1.0"
ffmpeg = "^1.4"
PyNaCl = "^1.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
